Improve handling of required shrink/stretch exceeding thresholds
When the required line shrinking needed to find a feasible solution is
less than the minimum threshold, give up and create an over-long line.

When the required line stretching needed to find a feasible solution is
more than the max threshold, try again with a higher threshold. What TeX
does here is first to apply hyphenation and retry with the same (or
similar) threshold and then retry with a higher threshold if that fails.
Hyphenation is not yet implemented in this library.