Fix handling of no lines with adjustment ratio in range
The logic for deciding whether to give up and add the current position
as a breakpoint or whether to increase the maximum allowed adjustment
ratio did not take account of possible _earlier_ breakpoints that were
ignored because the AR required exceeded the maximum.

Fix this by tracking the minimum AR needed above the threshold in order
to make layout succeed.