Implement hyphen handling in `positionItems`
 - Include the width of a penalty where a breakpoint occurs
   (typically the width of a hyphen) in the line width.
 - Include penalty items at the end of a line in the
   list of positioned items returned.
 - Rename the method to `positionItems` since it no longer returns
   just boxes.

A test is currently failing in a scenario where the algorithm choses a
line break at a point that is followed by another glue. This is
happening because `breakLines` does not ignore glue/penalty items at the
start of lines when calculating the total width/shrink/stretch for a
line and thus its adjustment ratio.