Do not include width of glue items at start of line in calculations
Given the sequence of items [box, glue, glue, box, penalty], if the first
glue is used as a breakpoint then we do not want to include the
shrink/stretch/width of the second glue when calculating the width of
the next line, since when rendering we skip glue items at the start or
end of lines.

This change produced some changes in the output of existing tests, which
have been adjusted accordingly.