Make `justifyContent` take an array of elements to avoid layout thrashing.
Previously calling `justifyContent` with a series of elements would
result in layout thrashing because the function first reads layout in
order to measure available space, then modifies the element's content to
insert linebreaks. Passing a list to `justifyContent` allows the all the
reads to be performed first, followed by all the writes.