Export hypher wrapper
The 'hypher' library is not currently packaged as a UMD or for direct
consumtion in the browser without jQuery. Since hyphenation is a
critical part of rendering justified text, even with paragraph-level
linebreaking, it makes sense to include it as a core feature.

Also create a UMD bundle for the en-us patterns so that tex-linebreak
can be conveniently used from the browser with a couple of script tags
using:

```html
<script src=".../dist/lib.js"></script>
<script src=".../dist/hyphens_en-us.js"></script>
<script>
const paragraphs = Array.from(document.querySelectorAll('p'));
const hyphenate = texLineBreak_lib.createHyphenator(texLineBreak_hyphens_en_us);
texLineBreak_lib.justifyContent(paragraphs, hyphenate);
</script>
```