Fix pathological performance issue in `breakLines`
When feasible breakpoints are found given the current max adjustment
ratio, the higher threshold used to retry was too low, possibly
resulting in a large number of retries before succeeding in being able
to process the whole paragraph.

The previous higher threshold for retrying was the _smallest_ ratio between a
pair of breakpoints considered in this iteration which was ignored because it
exceeded the current threshold. However this may be a much smaller value
than needed at the point in the paragraph where we failed to find a
break given the current adjustment ratio.

This commit uses double the minimum ratio so we'll find the necessary
break with far fewer iterations.